// packages/web/src/shared/api/models/songResponseDTO.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: v1
 */
import type { SongCategoryDTO } from './songCategoryDTO';
import type { SongLanguageDTO } from './songLanguageDTO';

export interface SongResponseDTO {
  /** @nullable */
  artist?: string | null;
  category?: SongCategoryDTO;
  difficulty?: number;
  /** @nullable */
  firstLine?: string | null;
  id?: number;
  language?: SongLanguageDTO;
  /** @nullable */
  spotifyId?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  year?: number | null;
}